requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.0.0
opencv-python>=4.7.0
tqdm>=4.65.0
lpips>=0.1.4
scikit-image>=0.20.0
scipy>=1.10.0
hydra-core>=1.2.0
omegaconf>=2.3.0
